# ChuckNorrisConnector/bin/Debug/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ChuckNorrisConnector_Connector",
    "version": "1.0"
  },
  "paths": {
    "/api/ChuckNorrisJokes/Subscribe": {
      "post": {
        "tags": [
          "ChuckNorrisJokes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SubscribeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ChuckNorrisJokes/Unsubscribe/{flowInstanceId}": {
      "delete": {
        "tags": [
          "ChuckNorrisJokes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "flowInstanceId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ChuckNorrisJokes/categories": {
      "get": {
        "tags": [
          "ChuckNorrisJokes"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ChuckNorrisJokes/joke": {
      "get": {
        "tags": [
          "ChuckNorrisJokes"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "SubscribeRequest": {
      "type": "object",
      "properties": {
        "flowName": {
          "type": "string"
        },
        "flowTriggerEndpoint": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  },
  "host": "gw6z58x5-7007.uks1.devtunnels.ms",
  "schemes": [
    "https"
  ]
}